challenge playing with bit operator

+++ section8/BiteShiftingChallenge.c
@@ -12,36 +12,36 @@ int decimalToBinary(int decimalNumber);
 
 int main(int argc, char const *argv[])
 {
-    long long number1;
-    long long number2;
-    long long  binaryNumber1;
-    long long  binaryNumber2;
+    int number1;
+    int number2;
+    long binaryNumber1;
+    long binaryNumber2;
 
     printf("Enter number 1: ");
-    scanf("%lli", &number1);
+    scanf("%i", &number1);
     
     printf("\nEnter number 2: ");
-    scanf("%lli", &number2);
+    scanf("%i", &number2);
 
     binaryNumber1 = decimalToBinary(number1);
     binaryNumber2 = decimalToBinary(number2);
 
-    printf("Number 1: %lli -> in binary is:%lli", number1, binaryNumber1);
-    printf("\nNumber 2: %lli -> in binary is:%lli", number2, binaryNumber2);
+    printf("Number 1: %i -> in binary is:%li", number1, binaryNumber1);
+    printf("\nNumber 2: %i -> in binary is:%li", number2, binaryNumber2);
 
-    long long complementNumber1 = ~number1;
-    long long complementNumber2 = ~number2;
+    long complementNumber1 = ~number1;
+    long complementNumber2 = ~number2;
 
-    long long orNumber = (number1 | number2);
-    long long andNumber = (number1 & number2);
-    long long xorNumber = (number1 ^ number2);
+    long orNumber = (number1 | number2);
+    long andNumber = (number1 & number2);
+    long xorNumber = (number1 ^ number2);
 
-    printf("\n%lli | %lli = %lli -> binary = %i", number1, number2, orNumber, decimalToBinary(orNumber));
-    printf("\n%lli & %lli = %lli -> binary = %i", number1, number2, andNumber, decimalToBinary(andNumber));
-    printf("\n%lli ^ %lli = %lli -> binary = %i", number1, number2, xorNumber,decimalToBinary(xorNumber));
+    printf("\n%i | %i = %li -> binary = %i", number1, number2, orNumber, decimalToBinary(orNumber));
+    printf("\n%i & %i = %li -> binary = %i", number1, number2, andNumber, decimalToBinary(andNumber));
+    printf("\n%i ^ %i = %li -> binary = %i", number1, number2, xorNumber,decimalToBinary(xorNumber));
 
-    printf("\nComplement number1 using (~) is:%lli", complementNumber1);
-    printf("\nComplement number1 using (~) is:%lli", complementNumber2);
+    printf("\nComplement number1 using (~) is:%li", complementNumber1);
+    printf("\nComplement number1 using (~) is:%li", complementNumber2);
 
 
     return 0;
